Fixed XYWing test.
The XYWing test was failing sometimes and other times succeeding. I
think this is because there were two possible XYWings and the randomness
in generating permutations from a set of cells meant that sometimes one
would be returned and some times the other.

This change effectively applies a naked pair and pointing pair before
checking for the XYWing (which would normally be found by the solver
first), leaving only the desired one to be found.