Added Rectangle Elimination strategy, fixed bug.
There is now a rectangle elimination strategy. Unlike the Python
version, it checks if both the wings are stronly linked to the hinge and
if so, places the value.

Fixed a bug with the XYZWing strategy: the cells that we could eliminate
from were not being scanned for val, and thus sometimes tried to
eliminate from a cell where val was not a candidate. This was
problematic because even though it wasn't actually eliminating anything,
it was not returning None so the solver would loop infinitely in such a
case.

Other minor changes add a few helpful methods, as well as implementing
some changes from clippy.