Start of AIC implementation.
Currently, aic.rs contains functions which make all the strong/weak
links and build all the AICs. It is not very good and was hacked
together as a proof of concept and isn't integrated into the solver.

Since the strong/weak link maps will also be used for coloring
strategies, that functionality should be separated and maybe the link
maps should be cached.