Fixed AIC bug, added name field to StrategyResult.
The AIC type 2 detection was not checking if the start and end of the
chain was in the same cell. This meant that invalid eliminations were
occuring, e.g. if both the start and end were in the same bivalue cell,
both candidates in it were being eliminated, leaving no candidates left
in that cell.

I also noticed that the detection of discontinuous AIC loops is missing
lots of cases - currently only a weak leak occuring in the end cell is
checked, but actually the candidate can be any that is weakly linked to
both ends. If it happens that the link between the discontinuity and the
start is also strong, then we can place the starting candidate. This
should be added.