Fixed Nice Loop/Type 1 AIC detection, AIC logic.
Discontinuous loop detection now uses the proper links to detect
discontinuities. Type 1 AIC check as its own function is also removed
because type 1 AICs are just a combination of discontinuous loops and so
are found that way.

Also now make use of the link maps that are already built instead of
calling Grid methods to find candidates to eliminate. This makes more
sense and simplifies the code.