Improved continuous AIC detection.
Continuous AICs were only being detected when the start and end
candidates were in the same cell. This does detect valid continuous
AICs, but will miss other valid loops. The proper criteria to use is
that the AIC is continuous if the start and end are weakly linked
(which, if they are in the same cell, is true). This change implements
that.