Added new type for AICs, changed build_aics fn.
The AICResult type is now used when an AIC that eliminates/places
candidates is found instead of StrategyResult. This is so the build_aics
fn can be used elsewhere (skyscraper, 2-string kite, turbot fish,
x-cycles, etc.) and then a StrategyResult can be built on top of that.

The build_aics fn now also doesn't store all AICs, but instead checks
each one as it encounters it. If it eliminates/places candidates, then
Some(AICResult) is returned. This means that only the needed AICs are
built, but does mean that they can't be sorted by length anymore.